+++ mayan/apps/tests/dependencies.py
@@ -22,11 +22,6 @@ PythonDependency(
     environment=environment_testing, module=__name__,
     name='django-test-migrations', version_string='==0.2.0'
 )
-PythonDependency(
-    environment=environment_testing,
-    module=__name__, name='django-test-without-migrations',
-    version_string='==0.6'
-)
 # Mock is set to production so that it is available in the Docker image
 # and allows running the test suit in production.
 PythonDependency(